Drift QueryParserError to implement errors adhering to the new spec
I aliased old QueryParserError to the QueryParserError::Legacy enum variant and adapt the code to operate on the old variant. Then I created a new QueryParserError::New variant for the new set of errors.

When all the errors get migrated to the new error format, we can remove the callsites  where Legacy was used and promote the `New` enum variant, converting again QueryParserError to an struct.