Use known errors for some of the core errors. Might break client. Read details
Before this change, all HandlerErrors were returned as unknown errors to the client, ie. No error code. Now, we reuse the core logic of rendering errors from the implementation of From<CoreError> for user_facing_error::Err, which can result in known errors for some of the types.

This might the correct thing to do, but too ahead in the refactoring. In a next commit I will only do that for validation errors.